requirements.txt
pydantic>=2.5,<3
pyyaml>=6.0,<7
click>=8.1,<9
rich>=13.0,<14
python-dateutil>=2.8,<3
